feat(canister_migration): [CON-1487] Add the nns delegation format to the execution_environment::QueryCache (#6461)
# Problem

In https://github.com/dfinity/ic/pull/6378 we are adding a
`/api/v3/canister/{effective_canister_id}/query` endpoint which differs
from the `/api/v2` endpoint in that we pass a different format of the
nns delegation to the canister. This means that depending on the
endpoint, the canister might produce a different response. Without
making any changes to the `execution_environment::QueryCache` it could
happen that a user sends the same query first to the `/api/v2` endpoint
and then to `/api/v3` which would return a cached `/api/v2` response
which could (if the canister chooses to) contain the certificate
delegation in unexepected format.

# Solution

To distinguish between cache entries we add
`CertificateDelegationFormat` to the key of the `QueryCache`.