feat: CON-1530 Introduce a way to order pre-signature stashes by their priority (lowest fill level) (#6464)
This PR introduces a struct that will allow us to order pre-signature
stashes by their priority. Specifically, stashes with a proportionally
lower fill level receive a higher priority. Ties between stashes with
the same proportional fill level are broken by comparing the key IDs of
the stashes instead.

This structure will allow us to determine for which stash the next
pre-signature creation should be started. In the future, consensus will
start the creation of new pre-signatures for the stash with the highest
priority (meaning the lowest fill level) until the maximum capacity for
creating more pre-signatures is reached.