chore(PocketIC): drop node_ids from SubnetConfig (#6468)
This PR drops `node_ids` from `SubnetConfig` returned by PocketIC as
part of its topology. The motivation for this change is that `node_ids`
are unlikely to be useful for PocketIC users unless they want to test
stuff related to node metrics in which case they likely need more
information anyway and thus retrieve the registry from PocketIC (e.g.,
from the state directory as shown in the test
`test_custom_blockmaker_metrics`).