chore(icrc-rosetta): deprecate network_type parameter [FI-1784] (#6434)
As a step towards deprecating the static testnet, the `network_type`,
where users can choose between `mainnet` and `testnet`, will be removed
in the future. Instead, users would be encouraged to develop locally, or
on mainnet and use our test faucet.

Once this is merged, the docs of Rosetta will be updated to not use
testnet or encourage it, and then we can proceed to remove this
parameter entirely.