test(consensus): Make query_stats_basic test more stable (#6479)
In the test, we update a universal canister to contain some data. Then
we query each node individually to return the data.

Occasionally, this fails, since a single node might be behind, and
therefore might serve outdated data.

This PR sets the waiting time from 2 secs to 5 secs, which hopefully
makes the test more robust.