fix: Some more cleanups in execution fuzzers (#6456)
The previous [PR](https://github.com/dfinity/ic/pull/6354) that cleaned
up the embedder flag for enabling memory64 missed a few things:

1. Setting up the test environment can be made easier in some execution
fuzz tests as we do not need to differentiate between an embedder that
allows 32-bit only or 64-bit Wasms (missed as there was no complaint
about the unused bool that controlled whether the embedder enables
`wasm64`).
2. While `SystemApiModule` was correctly producing both 32-bit and
64-bit Wasms with the respective imports of the system API,
`ICWasmModule` and `MaybeInvalidModule` were inadvertently changed to
only include the 64-bit version of the system API. Technically, this
works but it's better to keep the behaviour closer to what we had before
where for 32-bit Wasms we import the respective system API.

As a drive-by also renaming the boolean that used to control this from
`memory64_enabled` to `is_wasm64` to better reflect that it's not
related to enabling anything in the embedder but rather it controls the
type of the produced module.