feat: CON-1559 Purge pre-signatures that exceed the maximum stash size (#6504)
Under normal circumstances, consensus makes sure to never deliver any
pre-signatures to execution that would exceed the maximum configured
number of pre-signatures in the stash.

However, in case there is a bug in consensus, it would be bad to fill up
the pre-signature stash endlessly. Therefore, this PR additionally
enforces the maximum pre-signature stash size in the execution layer.
Any pre-signature exceeding the maximum size will be purged, and trigger
a warning + metric increase.

The same situation may occur naturally, if the maximum stash size was
reduced via NNS proposal. Before this PR, the pre-signature stash would
have remained at the previous configured size, until the entire stash is
purged and re-generated as part of the next key rotation. Now, the stash
will be trimmed to the new size immediately, thanks to the new logic.