test(nns): End-to-end NNS recovery test (#6407)
Design docs
([high-level](https://docs.google.com/document/d/1qguEp1aEMfuNFWTfZBCxuO-XrTxFKo6tmOdaqCqj5ao/edit?tab=t.0#heading=h.z87s6giniorx),
[more
detailed](https://docs.google.com/document/d/16WnpXebOrrdkoQscg7mRn2vUrAabbEN2Ijl2cR85GIs/edit?tab=t.0#heading=h.elr0a68qp5el))

This PR adds the missing functionality to `ic-recovery` for NNS recovery
to not rely on having `admin` SSH access on all nodes but on `backup`
SSH access on all nodes instead. This includes (below bullet points
mostly targeted to @consensus as code owner of `ic-recovery`):
- Introduce the `--backup-key-file` parameter that should contain the
SSH private key whose public key is setup to be accepted on all nodes
for the `backup` user. The `DownloadCertificationsStep` is thus also
adapted to use the `backup` user instead of `admin`.
- Note that `--key-file` is still required as `admin` access for a
remote recovery to be able to download the state and upload the
registry, CUP and state.
- Introduce the `--output-dir` parameter that will contain the
`recovery.tar.zst` and `recovery.tar.zst.sha256` files necessary for the
rest of the recovery.
- The `ICReplay` step replays until the highest certification height by
default.
- *Most interesting change*: The `UploadCUPAndRegistry` step no longer
uploads the CUP and registry (so-called “recovery artifacts”) to all
nodes, assuming `admin` SSH access on all nodes, but rather uploads them
to a single node (one for which we have `admin` SSH access). The other
nodes will be able to download those artifacts later thanks to the
`guestos-recovery-upgrader/engine`.
- Introduce a new `CreateArtifacts` step that creates the tarball
containing CUP and registry.

This PR also completes the end-to-end NNS recovery test. Overall
outline:
- Setup a functional 4-node NNS subnet with a virtualized HostOS.
- We unfortunately cannot setup a subnet like this directly, so we start
with a single node without a virtualized HostOS, propose to add 4 nodes
with a virtualized HostOS and finally propose to remove the original
one.
- Propose to add our SSH key as `backup` access on all nodes (which is
already the case in production).
- Break the subnet by replacing the replica binary with `/bin/false` on
`f+1` nodes.
- Run `ic-recovery` in remote-mode. This outputs `recovery.tar.zst` and
`recovery.tar.zst.sha256`.
- Host `recovery.tar.zst` on a HTTP server instantiated with a UVM. This
UVM takes the role of the upstreams in production
(download.dfinity.network & download.dfinity.systems).
- Also host the recovery GuestOS image there.
- Note: this image would normally contain the hash of `recovery.tar.zst`
in
`ic-os/components/misc/guestos-recovery/guestos-recovery-engine/expected_recovery_hash`,
but AFAIK we currently cannot build a GuestOS image at runtime, so we
decided to serve a default image and manually overwrite the file with
SSH in a later step.
- Simulate the action of node providers on `2f + 1` nodes in parallel:
- Reboot HostOS with `recovery=1` to trigger `guestos-recovery-upgrader`
which will download the recovery GuestOS image that contains
`guestos-recovery-engine`.
- [Done for testing but not in production] Spoof HostOS’s DNS to point
the upstreams to the UVM.
- Note: the UVM uses a TLS certificate signed by a CA that HostOS and
GuestOS trust. This allows to have a trusted TLS connection to the
upstreams without actually communicating with the production buckets.
    - HostOS will download the GuestOS image from the UVM and launch it.
- [Done for testing but not in production] As described earlier, it is
challenging to build a GuestOS image at runtime, so we overwrite the
expected hash by SSHing in the GuestOS.
- [Done for testing but not in production] Spoof GuestOS’s DNS to point
the upstreams to the UVM.
- Note: the UVM uses a TLS certificate signed by a CA that HostOS and
GuestOS trust. This allows to have a trusted TLS connection to the
upstreams without actually communicating with the production buckets.
- GuestOS runs `guestos-recovery-engine`, which downloads the recovery
artifacts from the UVM and extracts them in the correct locations,
effectively accomplishing the old functionality of the
`UploadCUPAndRegistry` step in `ic-recovery` which required `admin` SSH
access.
- Ensure that the subnet is healthy again (i.e. responds to query and
update calls)

Further work: We will add tests for more recovery scenarios. Examples
include running a local recovery instead of a remote one, include the
DFINITY-owned node in the `f+1` broken nodes, break all nodes instead of
`f+1`, break the orchestrator binary instead of the replica.

Note for @node: I think it would be best to refactor
`rs/recovery/src/lib.rs` into different files, especially because this
test has quite a different setup than the others (multiple nodes, custom
DKG interval, setup an additional UVM), but I see that you already have
a ticket for that:
[NODE-1642](https://dfinity.atlassian.net/browse/NODE-1642).

Note for @idx: It is not yet clear to me how we would build and upload
the recovery GuestOS image in production. Andrew suggested to use
`//ic-os/guestos/envs/recovery:bundle-update` to build it, is that
sufficient? Do we need to commit the expected hash of the recovery
artifacts to the repo and if yes, on what branch? What about uploading
the image? I guess we could use `//ci/src/artifacts:upload.sh`, but do
we need to modify the current script in any way?
I think we could schedule a meeting to discuss this in more detail.


[NODE-1642]:
https://dfinity.atlassian.net/browse/NODE-1642?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ