feat: Start IDkgBlockReader above certified height during payload creation & validation (#6308)
In the future, we want to start storing a large number of pre-signatures
in the replicated state, up to a maximum.

Due to the asynchrony of execution and consensus, determining the total
number of existing pre-signatures requires some care: We need to count
the number of pre-signatures that are currently part of the certified
state, in addition to all pre-signatures being delivered in subsequent
blocks that have not been executed/certified yet.

This PR ensures that during payload creation and validation, the
`IDkgBlockReader` starts at height `min(summary_height,
certified_height+1)`. This allows us to calculate the total number of
pre-signatures in the blockchain above the certified height.

The approach is similar to what is done in the batch payload builder,
which collects all past payloads above the certified height during each
payload
[validation](https://sourcegraph.com/github.com/dfinity/ic@12f93cd3f38337d6a4b98272d360990f808221cf/-/blob/rs/consensus/src/consensus/validator.rs?L1260-1264)
and
[creation](https://sourcegraph.com/github.com/dfinity/ic@12f93cd3f38337d6a4b98272d360990f808221cf/-/blob/rs/consensus/src/consensus/block_maker.rs?L422-423).

Note that generally, blocks below the summary are not guaranteed to
exist, because they are purged once the CUP exists. However, if we have
such a CUP, that implies there is already a finalized block `b` with
`b.certified_height >= summary_height`, which means we should not be
creating/validating any blocks referencing a lower certified height
manually. Such blocks should instead be validated using the notarization
fast-path.