chore: Stop preparing routing table in the old format outside of production (#6059)
The old routing table format is no longer read by the clients, and it's
no longer validated by the registry client in the previous PR (#6034).
This PR stops preparing the routing table with the "routing_table" key.
This cannot be done at the same time as the previous PR because some
tests run by CI uses the mainnet registry canister, so they won't pass
until the previous PR gets into the mainnet.