feat(ICP_Ledger): FI-1727: Trap in case the deprecated ICP ledger notify method is called (#5734)
Trap in case the deprecated `notify` method of the ICP ledger is called.
The trap message include a pointer to [the forum
post](https://forum.dfinity.org/t/deprecating-the-ledger-notify-flow-for-minting-cycles-in-favor-of-cmc-notify/42502)
with the instructions on how to migrate to the new CMC notify flow.

Before proposing the changes to the behavior of the `notify` method in
this PR, the following steps were taken:
- The deprecation of the `notify` method was
[announced](https://forum.dfinity.org/t/deprecating-the-ledger-notify-flow-for-minting-cycles-in-favor-of-cmc-notify/42502)
on the forum in March, with a follow-up in May.
- [Logging](https://github.com/dfinity/ic/pull/5096) of the callers of
the `notify` method was added, and it was determined that since the
introduction of the logging a month ago, the method was called by a
canister from the SONIC SNS, and a handful of other non-canister
principals. The SONIC team was explicitly informed about the upcoming
change.
- We did consider a fail-with-probability approach, but decided against
it.

---------

Co-authored-by: Daniel Wong <97631336+daniel-wong-dfinity-org@users.noreply.github.com>
Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>