fix: the bazel-bep artifact upload of the guest-os-qualification job (#6519)
We recently had the Release Testing workflow fail multiple times because
the
[guest-os-qualification](https://github.com/dfinity/ic/blob/69b755062f5ef0a7d6efc9a127172b46121420c8/.github/workflows/release-testing.yml#L119)
job was run on two versions instead of one causing the [Upload
bazel-bep](https://github.com/dfinity/ic/blob/master/.github/actions/bazel/action.yaml#L71)
step to fail since it already uploaded an artifact with the same name
for the other IC version.

To fix this we now include the `matrix.version` in the name of the
bazel-bep artifact to upload.

---------

Co-authored-by: IDX GitHub Automation <>