test(MR): Speed up rejoin_large_test (#6482)
While taking a look at the state sync system tests, I came along this
test, which seems to be similar to what I want to use for the fast state
sync benchmark.

However, it seems the test was running much slower than needed.
This PR:

- Parallelizes the calls inside the `modify_canister_heap`, such that
the different canisters run in parallel. This speeds up the time to fill
them up from 150s to around 40s.
- Sets a shorter DKG interval, from 500 to 25. A large amount of time is
spend waiting for cups
- The unit and initial notary delays are set to 200ms