feat(ICRC-Ledger): FI-1653: Ensure upgrade u64 <-> u256 fails (#6486)
We add `token_type` equal to `U64` or `U256` to the ledger state. During
`post_upgrade` we check if the deserialized `token_type` is the same as
the token type of the current wasm.
If an older ledger is upgraded and the `token_type` cannot be checked,
we additionally read the first balance in `post_upgrade`. The wasms use
incompatible encodings for storing numbers in stable structures and the
read will fail if a u64 ledger is upgraded with a u256 wasm. If the
upgrade was successful, the ledger would later fail when trying to read
the balances map.