feat(PocketIC): enable HTTP gateway creation together with instance creation (#6469)
This PR extends the argument of the endpoint `/instances/` with an
additional optional field `http_gateway_config` specifying that an HTTP
gateway with the given configuration should be created for the newly
created instance.

The motivation for this change is bootstrapping frontend canisters as
part of `IcpFeatures`: some frontend canisters, e.g., the NNS dapp,
require a gateway URL in its initial argument at deployment.