chore: Add exclusive filtering to MixedHashTree
Previously, commit 09a91114ebaf633d7a8dd9785838ba673533eae4 introduced a
way to prune a `MixedHashTree` further by providing an inclusive list of
paths to keep, and anything not in the filter would be pruned. This
functionality was then used to implement new HTTP endpoints in commit
cb4878c4688c1e2d669f6245ff0e33c78660c57e.

This commit now provides a variant API `pruned`, where the input is instead
an exclusive list of paths to prune, and any path not mentioned in the
input is kept instead. It is intended as a more ergonomic helper
function in some use cases and will be used for upcoming changes to the new
HTTP endpoints.