feat(ICRC-Archive): FI-1844: Ensure upgrade u64 <-> u256 fails (#6546)
We add an additional field in the archive config: `token_type`. This
field is set to the token type supported by the archive wasm (`U64` or
`U256`) and checked during the upgrade. Since current mainnet archive
wasms don't have this field yet, this check will only be active after
the next upgrade of mainnet wasms (i.e. an upgrade from a version
containing this PR to a later version).