fix(orchestrator): Introduce timeout with backoff when fetching CUPs (#6219)
When streaming the response body of the CUP endpoint, we currently rely
on the 15 min global timeout of the upgrade loop, in case the server
stops sending data.

This PR adds a dedicated timeout mechanism for streaming the CUP body.
Initially, the timeout is set to 30 seconds, and is multiplied by a
factor of two whenever it is hit. The timeout is reset to 30 seconds if
the body can be streamed successfully in its entirety. Assuming the
bandwidth requirement of 300 mbit/s is met, this should still allow
nodes to fetch CUPs of a size above 1GB in the first try.

Additionally, this PR reverts a (likely accidental) change in
https://github.com/dfinity/ic/commit/a9d5be10daea5b19fae4aea23e55126ab4a7d07b,
and again makes the orchestrator call the CUP endpoint of the local
replica first, before trying the replica of a different peer. This
should reduce upgrade downtime of some nodes, and reduce the likelihood
of hitting a bad node.

Finally, we change the signatures of some functions from `fn foo(log:
&Logger) -> Option<T>` to `fn foo() -> Result<T, String>`, which
improves maintainability.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>