chore(PocketIC): use Option<EmptyConfig> instead of bool and Option<bool> (#6549)
This PR replaces `bool` and `Option<bool>` by `Option<EmptyConfig>` in
public PocketIC types for the sake of backward-compatibility if we need
to add further parameters in the future.

The original values `false`, `None`, and `Some(false)` correspond to
`None` while the original values `true` and `Some(true)` correspond to
`Some(EmptyConfig {})`.