chore(fuzzing): Make sandbox binary identifiable via a custom section (#6568)
Right now, for a binary to act as a sandbox executable, it's name must
be included in `RUNNABLE_AS_SANDBOX`. However, this is not flexible if
we need to include new binaries as this must be modified every time.

Instead, the library exposes a macro which embeds a custom section in
the target binary which is checked by the launcher to identify the
target binary as a sandbox. This code is gated with `#[cfg(feature =
"fuzzing_code")]`, so it shouldn't block the release code path.