feat: CON-1530 Start the creation of new pre-signatures for the emptiest stash (#6477)
This PR adds a new function that, given the total amount of existing
pre-signatures, starts the creation of new pre-signatures as part of the
current payload.

New pre-signatures are started iteratively for the currently emptiest
stash. This is done using a priority queue, which is initialised with
the fill levels of all stashes, and updated whenever a new pre-signature
is started.

The following constraints apply:
- The next pre-signature is started for the currently emptiest stash
- The maximum size of a stash may not be exceeded
- There is a limit to the number of pre-signature transcripts we want to
work on in parallel. Once this limit is reached, no additional
pre-signatures should be started

The capacity of how many pre-signature transcripts should be worked on
in parallel for any given payload is controlled by the new registry
parameter `max_parallel_pre_signature_transcripts_in_creation`. Each
pre-signature in creation consumes some of this capacity. Schnorr
pre-signatures consist of a single transcript, and therefore consume 1
transcript in capacity. ECDSA pre-signatures require working on two
transcripts in parallel, and therefore consume 2 capacity points.

For this reason the following situation may happen: The emptiest stash
asks for an ECDSA pre-signature, however, there is only enough capacity
to create a Schnorr pre-signature. In that case, we should _not_ start a
new Schnorr pre-signature and instead wait until enough capacity exists
in the payload to start the creation of a new ECDSA pre-signature which
has the highest priority. This is to prevent the creation of ECDSA
pre-signatures being starved.