chore: CON-1583 Add a topology version metric and use it in some tests adding nodes (#6574)
This PR adds two metrics:
- the registry version of the latest subnet topology relevant to P2P
- the registry version of the earliest subnet topology relevant to P2P
(this is equivalent to the "oldest registry version in use" by
consensus).

Additionally, we use the latter metric to improve the robustness of two
system tests (`adding_nodes_to_subnet` and `tecdsa_add_nodes`):

These tests execute proposals to add nodes to a subnet, and then assert
that the added nodes make progress as part of the new subnet.
Technically, this isn't enough to ensure that the new nodes indeed
joined the _active_ subnet membership, and participate in consensus
after a successful resharing of the old membership. The reason is that
such an assertion will always pass as long as the new nodes manage to
sync blocks, even if consensus is still driven by the old membership.

Therefore, this PR adds a check making sure that the "earliest subnet
topology" used by P2P progresses as expected with each membership
change. This indicates a successful resharing of (N)IDkg transcripts to
the new membership, and therefore the release of the old membership.

An alternative approach would have been the following: After adding all
nodes to the subnet, kill f+1 nodes of the new membership. If the subnet
stalls, then these nodes did indeed join the active subnet membership.
This is done by some other tests around membership changes such as
`network_large_test` and `node_assign_test`.