feat(proposal-cli): create forum topic for canister upgrade proposals (#6571)
As part of the Voting Neuron Grants season 2 teams submitting canister
upgrade proposals are required to [create a forum
topic](https://forum.dfinity.org/t/nns-proposal-discussions/34492) for
every proposal or group of proposals. This PR adds a new functionality
to `proposal-cli` to be able to create a new forum topic given one or
several proposal IDs.