fix(nns-recovery): do not fix DFINITY-owned node twice (#6606)
The NNS recovery test is flaky. #6554 fixed a first edge-case and this
one should fix a second one.

In some cases, the DFINITY-owned node would be fixed twice: once by
`ic-recovery` and once by the
`guestos-recovery-upgrader`/`guestos-recovery-engine`. This is not a
problem per se, the bug is a "test" bug: when simulating the actions of
node providers, we overwrite `BOOT_ARGS_A` in `/boot/boot_args`
[here](https://github.com/dfinity/ic/pull/6606/files#diff-c26ca29faacddc5919f46b7b6d2d7c503af940fe2e7b14038964accb17d0bebdL256).
This works fine if the node is currently using partition A. But the
DFINITY-owned node has already upgraded as part of `ic-recovery` and
thus is using partition B. This led to its state being wiped and not be
able to make consensus progress.

As a short-term flakiness fix, this PR does not run the simulated NP
actions on the DFINITY-owned node anymore, but the functionality can be
introduced more carefully in the future as part of the effort of testing
more recovery scenarios.