fix(recovery): escape ic-recovery filepaths (#6613)
The code breaks if the path to the key file or output directory have a
space in them.

A concrete example is by running a system test with a space inside its
name. For example, launching a system test with
`.add_test(systest!(nns_recovery_test; TestConfig {}))` (which will be
the format for NNS recovery tests) gives the path
`…/tests/nns_recovery_test (TestConfig
{})/ssh/authorized_priv_keys/admin`.

This PR properly escapes the paths (I tested using the path above, and
it now works).