feat: adding node states to the orchestrator to track state changes (#6572)
This feature adds the possibility for the orchestrator to know when the
node is leaving a subnet it was assigned to.

For now it is needed to display certain messages on the BMC console if
the node is leaving the subnet and that the node shouldn't be turned off
just yet. Later when the node leaves, it will display a different
message that it is now safe to turn it off.

Supporting pictures taken from kibana for graceful leaving:
<img width="1528" height="577" alt="image"
src="https://github.com/user-attachments/assets/45e83949-6ca8-4eb4-a805-6359e40dd20d"
/>