feat: [MR-715] [MR-382] Skip applying chunks from file group chunk if they are already copied (#6625)
The metric state_sync_chunks_remaining is keeping track of the number of
chunks that have not been downloaded or copied yet.

However, there are chunks that can potentially be both copied and
downloaded in the same state sync. If a chunk is part of a file group,
we download them regardless of whether they have been copied before.

This PR fixes the issue by skipping applying chunks if they are already
copied. We still download them though to simply the logic of downloading
file group chunks.

In addition to fixing the metrics, this PR is a prerequisite for
hardlinking files during state sync as we cannot hardlink and then
overwrite. This PR adds a test creating a scenario where individual
chunks from file groups chunks are both copied and downloaded and we
verify they are properly skipped by checking the file metadata.

---------

Co-authored-by: Stefan Schneider <31004026+schneiderstefan@users.noreply.github.com>