fix(guestos-recovery-upgrader): curl fails on server errors (#6652)
This PR adds the `--fail` flag to `curl`, similar to the
[guestos-recovery-engine](https://github.com/dfinity/ic/blob/d0996890b8edd4af8923deb6607b55f80f587950/ic-os/components/misc/guestos-recovery/guestos-recovery-engine/guestos-recovery-engine.sh#L43),
to make curl fail on server errors, like a 404 if the recovery image is
not present. This correctly attempts the download on the other upstream,
and does not try to compare the hash of the 404 XML/HTML page returned
by S3/R2, that yields confusing logs otherwise.

Please let me know if this was actually intended.