feat(ICRC-Rosetta): FI-1846: Handle burn and mint fees (#6620)
For burn blocks, if the fee is set, it is deducted from the `from`
account, and credited to the fee collector, if set.
For mint blocks, if the fee is set, it is deducted from the amount
credited to the `to` account, and credited to the fee collector, if set.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>