feat: CON-1533 Introduce a Consensus flag for storing pre-signatures in state (#6529)
This PR introduces a feature flag in consensus for storing
pre-signatures in the replicated state. If pre-signatures are stored in
the replicated state, consensus behaves differently as follows:

### 1. Pre-signature storage
Available pre-signatures are no longer maintained on the blockchain.
That means every available pre-signature is part of exactly one block
(the block that delivers it to execution), before it is purged from the
blockchain. Previously, available pre-signatures remained on the
blockchain until a response for the request that the pre-signature was
paired with, was generated.

### 2. Pre-signature creation
If pre-signatures are stored in the replicated state, then consensus
starts the generation of new pre-signatures based on the capacity in the
current payload, and the space in the pre-signature stash. Previously,
only the payload capacity was considered.

### 3. Signature expiration
If pre-signatures are stored in the replicated state, then signature
requests may time out even if they haven't been matched to a
pre-signature yet. Previously, this was not possible for the following
reason:

Consider an unmatched request context at certified state height 10. When
creating an IDkgPayload at height 12 that references the certified state
at height 10, we may see the unmatched request context and decide to
generate a reject response for it, because it expired. However, in that
case the request context may still be paired with a pre-signature at
height 11, before the reject response is delivered which would remove
the context. This pre-signature subsequently would not be removed from
the payload, and then paired with a second request context, which should
never happen.

This is not a problem in the new implementation, because pre-signatures
are consumed from the pre-signature stash after pairing, so there is no
way for it to be matched to a second request.

## Upgrade compatibility
### Upgrade (enable flag)
All existing available pre-signatures on the blockchain are removed in
the first payload created by the new version. However, ongoing signature
request that were already paired with pre-signatures of the old version
can still be completed, as these pre-signatures are additionally part of
the request context in replicated state. Newly create pre-signatures
will be stored in the pre-signature stash, where they will remain until
they are consumed by signature requests.

### Downgrade (disable flag)
There should be no available pre-signatures on the blockchain after the
downgrade, because summary blocks created by the version with the flag
enabled don't contain available pre-signatures. During the first
execution round, the pre-signature stash is purged. Ongoing signature
request that were already paired with pre-signatures of the old version
can still be completed, as the pre-signatures are part of the request
context in replicated state. Newly created pre-signatures will once
again be maintained on the blockchain.

## Testing
The feature flag will remain disabled for now. A PR enabling it (and
running all system tests) can be found here:
https://github.com/dfinity/ic/pull/6530