chore(nns): Small cleanups related to topics and NNS functions (#6679)
# Why

Before introducing `committed_topics` in to known neuron data, clean up
the existing `Topic` in the API type as it's actually not used, so that
there is less confusion when adding a `Topic` type in the API.


# What
Some cleanups related to topics and NNS functions:

- NetworkCanisterManagement->ApplicationCanisterManagement. This change
is entirely internal, and externally the topic 8 is already interpreted
as ApplicationCanisterManagement
- Remove the as_str_name and from_str_name for Topic API type. They were
generated by prost but they are no longer prost types.
- No conversion between `pb_api::Topic and pb::Topic`, they are never
needed.
- Call `compute_topic_at_creation` once after `validate_proposal`,
simplifying the reasoning that the topic isn't `Unspecified`
- Change `compute_topic_at_creation` to return error instead of using
`Topic::Unspecified` to signal error.
- Obsolete NNS functions now returns `Unspecified` topic
- Backward compatibility concern: the compute_topic_at_creation is only
called at proposal creation time