test: [MR-507] Make state sync tests less compressible (#6682)
Currently, the canister state in statesync_test canister is sparese.
During network transmission, the chunks are compressed by 200-500x.
In order to accurately test the state sync performance in the
downloading phase under certain network settings(e.g. in
https://github.com/dfinity/ic/pull/6657), we should make the chunks less
compressible.
This PR is a simple solution to use `SmallRng` to generate random data
on the whole state. The compression ratio now is virtually 1, which is
great for estimating downloading speed. I
t only adds a few seconds in expanding or modifying 8GiB of data during
the test. In the future, for tests with larger state(e.g. 1 TiB), we
could use some other ways, such as shifting to create similar but
different incompressible chunks.