chore(p2p): Add metrics for the number of concurrent incoming/outgoing QUIC streams (#6749)
For each artifact to be transmitted via a connection between two peers,
P2P opens a new and dedicated QUIC stream. The maximum number of
concurrent QUIC streams (and therefore the maximum number of artifacts
to be transferred in parallel) is controlled by the constant
[MAX_CONCURRENT_BIDI_STREAMS](https://sourcegraph.com/github.com/dfinity/ic@1ad8f453823f58e9fd48e251fb7fe30b88a56c89/-/blob/rs/p2p/quic_transport/src/connection_manager.rs?L74).

This PR adds metrics for the number of concurrent incoming/outgoing QUIC
streams, which will help us determine how far away we are from reaching
the configured maximum. This is especially relevant for IDKG, which
transmits a lot of artifacts.