chore: [NODE-1708] Update matching version assert (#6677)
After https://github.com/dfinity/ic/pull/6639, we can use
`without_unassigned_config()` when creating a subnet to skip setting the
unassigned node config to the deployed version. This can be used to
avoid the "extra upgrade" issue in tests when the initial GuestOS and
SetupOS/HostOS versions do not match.