feat: Allow more complex `read_certified_state` calls (#6771)
This commit extends `StateReader` with the function
`read_certified_state_with_exclusions`, allowing requests that take both
an list of included paths and excluded paths. This has no effect on the
public API for `read_state`, which only takes a list of paths to
include.

The new, more complex patterns are useful in cases where large subtrees
are requested, but specific leafs in the subtree should be omitted.

---------

Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>