feat(bitcoin): add dogecoin header validation to adapter (#6644)
[XC-422](https://dfinity.atlassian.net/browse/XC-422?atlOrigin=eyJpIjoiZmZiYzRhNjU0NTMxNGZjNjk4YTk4MWViYzU4OGViZmQiLCJwIjoiaiJ9):
This PR adds dogecoin header validation into the adapter. The code is
mostly copy pasted from the
[ic-doge-validation](https://github.com/dfinity/dogecoin-canister/tree/master/validation)
in the Dogecoin canister but has been adapted to be consistent with the
checks done by the Bitcoin adapter. In particular, the differences are
as follows (see
[XC-485](https://dfinity.atlassian.net/browse/XC-485?atlOrigin=eyJpIjoiYzc3NzZmMDhhYjc3NDM0MWIyNDAwMGE3MjM3NzZiNTYiLCJwIjoiaiJ9)):
- Adapter but not canister checks headers against a list of hard-coded
checkpoints (taken from Bitcoin/Dogecoin core).
- Adapter but not canister checks that the header is within one year
from the tip.
- Canister but not adapter checks that the header is less than 2 hours
in the future.

Additionally, the generic type `Header` has been removed from the
`HeaderStore` trait. Indeed, `HeaderStore` do not need to contain the
AuxPow information of previous headers to validate an AuxPow header,
hence storing normal headers is enough.

Note: this PR does not touch Bitcoin header validation at all. In a
subsequent PR, it could be possible to refactor Bitcoin header
validation to use the parametrized methods and traits introduced in this
PR (similar to
[ic-doge-validation](https://github.com/dfinity/dogecoin-canister/tree/master/validation)).

[XC-422]:
https://dfinity.atlassian.net/browse/XC-422?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[XC-485]:
https://dfinity.atlassian.net/browse/XC-485?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ